+++ r-alert-activity-log.tf
@@ -1,34 +1,7 @@
-resource "azurerm_monitor_action_group" "action_group_notification" {
-  name                = local.action_group_name
-  resource_group_name = var.resource_group_name
-  short_name          = var.action_group_short_name
-
-  dynamic "webhook_receiver" {
-    for_each = var.action_group_webhooks
-
-    content {
-      name                    = webhook_receiver.key
-      service_uri             = webhook_receiver.value
-      use_common_alert_schema = true
-    }
-  }
-
-  dynamic "email_receiver" {
-    for_each = var.action_group_emails
-    content {
-      name                    = email_receiver.key
-      email_address           = email_receiver.value
-      use_common_alert_schema = true
-    }
-  }
-
-  tags = merge(local.default_tags, var.extra_tags)
-}
-
 resource "azurerm_monitor_activity_log_alert" "activity_log_alert" {
   for_each = var.activity_log_alerts
 
-  name        = coalesce(each.value.custom_name, data.azurecaf_name.alert[each.key].result)
+  name        = coalesce(each.value.custom_name, data.azurecaf_name.activity_log_alerts[each.key].result)
   description = each.value.description
 
   resource_group_name = coalesce(each.value.resource_group_name, var.resource_group_name)