+++ r-naming.tf
@@ -9,13 +9,26 @@ data "azurecaf_name" "action_group" {
 }
 
 # Custom naming until managed by CAF provider
-data "azurecaf_name" "alert" {
+data "azurecaf_name" "activity_log_alerts" {
   for_each = var.activity_log_alerts
 
   name          = var.stack
   resource_type = "azurerm_resource_group"
-  prefixes      = compact([local.name_prefix, var.use_caf_naming ? "alert" : ""])
-  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, each.key, var.use_caf_naming ? "" : "alert"])
+  prefixes      = compact([local.name_prefix, var.use_caf_naming ? "mal" : ""])
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, each.key, var.use_caf_naming ? "" : "activity_log_alert"])
+  use_slug      = false # var.use_caf_naming
+  clean_input   = true
+  separator     = "-"
+}
+
+# Custom naming until managed by CAF provider
+data "azurecaf_name" "metric_alerts" {
+  for_each = var.metric_alerts
+
+  name          = var.stack
+  resource_type = "azurerm_resource_group"
+  prefixes      = compact([local.name_prefix, var.use_caf_naming ? "mma" : ""])
+  suffixes      = compact([var.client_name, var.location_short, var.environment, local.name_suffix, each.key, var.use_caf_naming ? "" : "metric_alert"])
   use_slug      = false # var.use_caf_naming
   clean_input   = true
   separator     = "-"